# docs/openapi-spec.yaml
openapi: 3.0.0
info:
  title: task-management
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
  /v1/comments/{id}:
    get:
      tags:
        - v1 > comments
      summary: get by id
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
          example: '1'
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249868200216'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '308'
            ETag:
              schema:
                type: string
                example: W/"134-XvzpuC659HMcXgn+GOOypR1pFkA"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 02:55:36 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                error: null
                data:
                  id: 1
                  userId: 1
                  taskId: 5
                  content: this is my comment
                  createdAt: '2024-08-31T12:27:33.742Z'
                  updatedAt: '2024-08-31T12:29:19.662Z'
                  deletedAt: null
                  user:
                    id: 1
                    username: user1
                    createdAt: '2024-08-31T08:11:24.301Z'
                    updatedAt: '2024-08-31T08:11:24.301Z'
                    deletedAt: null
        '404':
          description: Not Found
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249868199938'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '33'
            ETag:
              schema:
                type: string
                example: W/"21-hExXrc+B/JrRtNZvDXrmSmOg6/k"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 03:00:13 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                data: null
                error: Not Found
    put:
      tags:
        - v1 > comments
      summary: replace a comment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                userId: 2
                taskId: 6
                content: Hello World
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
          example: '11111'
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17250084958398'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '301'
            ETag:
              schema:
                type: string
                example: W/"12d-CUcNDASUfwS87GMc6o41GYWQlC8"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 08:56:53 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                error: null
                data:
                  id: 1
                  userId: 2
                  taskId: 6
                  content: Hello World
                  createdAt: '2024-08-31T12:27:33.742Z'
                  updatedAt: '2024-09-01T08:56:53.962Z'
                  deletedAt: null
                  user:
                    id: 2
                    username: user2
                    createdAt: '2024-08-31T08:11:24.301Z'
                    updatedAt: '2024-08-31T08:11:24.301Z'
                    deletedAt: null
        '400':
          description: Bad Request
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249870295981'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '38'
            ETag:
              schema:
                type: string
                example: W/"26-zAb7KIIX9xyOeTaPku6Iz/KxjUw"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 03:01:10 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                data: null
                error: invalid taskId
        '404':
          description: Not Found
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249870295962'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '33'
            ETag:
              schema:
                type: string
                example: W/"21-hExXrc+B/JrRtNZvDXrmSmOg6/k"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 03:01:29 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                data: null
                error: Not Found
    patch:
      tags:
        - v1 > comments
      summary: update a comment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                userId: 2
                taskId: 6
                content: Hello World
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
          example: '1'
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17250082146920'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '302'
            ETag:
              schema:
                type: string
                example: W/"12e-Xo90VCE9s0AEThnpvxopop6itU0"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 08:52:31 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                error: null
                data:
                  id: 1
                  userId: 2
                  taskId: 6
                  content: Hello World2
                  createdAt: '2024-08-31T12:27:33.742Z'
                  updatedAt: '2024-09-01T08:52:31.503Z'
                  deletedAt: null
                  user:
                    id: 2
                    username: user2
                    createdAt: '2024-08-31T08:11:24.301Z'
                    updatedAt: '2024-08-31T08:11:24.301Z'
                    deletedAt: null
    delete:
      tags:
        - v1 > comments
      summary: delete comment
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
          example: '1'
      responses:
        '204':
          description: No Content
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249870806015'
            ETag:
              schema:
                type: string
                example: W/"e-hDOZGZ11/MTPwZq/A8sGFlQzkPc"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 02:59:56 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            text/plain:
              schema:
                type: string
              example: null
  /v1/comments:
    post:
      tags:
        - v1 > comments
      summary: create a comment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                userId: 1111
                taskId: 6
                content: this is my comment
      responses:
        '201':
          description: Created
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249869445731'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '308'
            ETag:
              schema:
                type: string
                example: W/"134-ol4ztlnhywJKDFB1UOqJMtNn1wk"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 02:57:40 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                error: null
                data:
                  id: 2
                  userId: 1
                  taskId: 6
                  content: this is my comment
                  createdAt: '2024-09-01T02:57:40.527Z'
                  updatedAt: '2024-09-01T02:57:40.527Z'
                  deletedAt: null
                  user:
                    id: 1
                    username: user1
                    createdAt: '2024-08-31T08:11:24.301Z'
                    updatedAt: '2024-08-31T08:11:24.301Z'
                    deletedAt: null
        '400':
          description: Bad Request
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17249869445549'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '38'
            ETag:
              schema:
                type: string
                example: W/"26-Lk0aj2H5f+q8gqkA5E/4mgrpJC0"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 03:00:42 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                data: null
                error: invalid userId
    get:
      tags:
        - v1 > comments
      summary: query
      parameters:
        - name: pagination[page]
          in: query
          schema:
            type: integer
          example: '1'
        - name: pagination[pageSize]
          in: query
          schema:
            type: integer
          example: '10'
        - name: filters[userId]
          in: query
          schema:
            type: integer
          example: '1'
        - name: sort[0]
          in: query
          schema:
            type: string
          example: createdAt:asc
        - name: filters[createdAtStart]
          in: query
          schema:
            type: string
          example: '2024-08-31'
        - name: filters[createdAtEnd]
          in: query
          schema:
            type: string
          example: '2024-09-30'
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            X-RateLimit-Limit:
              schema:
                type: integer
                example: '100'
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: '99'
            X-RateLimit-Reset:
              schema:
                type: integer
                example: '17250080151013'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '372'
            ETag:
              schema:
                type: string
                example: W/"174-2QN/B5gJEdNwX59sPWDBfwIkPoQ"
            Date:
              schema:
                type: string
                example: Sun, 01 Sep 2024 08:49:08 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                error: null
                data:
                  - id: 2
                    userId: 1
                    taskId: 6
                    content: this is my comment
                    createdAt: '2024-09-01T02:57:40.527Z'
                    updatedAt: '2024-09-01T02:57:40.527Z'
                    deletedAt: null
                    user:
                      id: 1
                      username: user1
                      createdAt: '2024-08-31T08:11:24.301Z'
                      updatedAt: '2024-08-31T08:11:24.301Z'
                      deletedAt: null
                pagination:
                  total: 1
                  pageCount: 1
                  page: 1
                  pageSize: 10
...